Moving character offset to the feet
**Important change!**

The Woka "origin" is now set at the bottom of the sprite (16, 32), instead of the center of the sprite.
This makes more sense as we consider someone enters a zone when the FEET actually enter the zone (and not the center of the body).

Though the new behaviour is the expected behaviour by most users, this will impact zone entry detection and scripting API position fetching.